+++ src/Organizations/constants.js
@@ -11,6 +11,7 @@ export const ORGANIZATION_SECTIONS = {
   accountsSection: 'accountsSection',
   notesSection: 'notesSection',
   integrationDetailsSection: 'integrationDetailsSection',
+  agreements: 'linkedAgreements',
 };
 
 export const ORGANIZATION_SECTION_LABELS = {
@@ -22,10 +23,11 @@ export const ORGANIZATION_SECTION_LABELS = {
   [ORGANIZATION_SECTIONS.vendorTermsSection]: <FormattedMessage id="ui-organizations.vendorTerms" />,
   [ORGANIZATION_SECTIONS.accountsSection]: <FormattedMessage id="ui-organizations.accounts" />,
   [ORGANIZATION_SECTIONS.integrationDetailsSection]: <FormattedMessage id="ui-organizations.integrationDetails" />,
+  [ORGANIZATION_SECTIONS.agreements]: <FormattedMessage id="ui-organizations.linkedAgreements.section" />,
 };
 
-export const CREATE_UNITS_PERM = 'organizations.acquisitions-units-assignments.assign';
-export const MANAGE_UNITS_PERM = 'organizations.acquisitions-units-assignments.manage';
+export const CREATE_UNITS_PERM = 'ui-organizations.acqUnits.assign';
+export const MANAGE_UNITS_PERM = 'ui-organizations.acqUnits.manage';
 
 // Mapping between attribute (field) in form and id of accordion
 export const MAP_FIELD_ACCORDION = {